Turn NetworkExploration into an interface
Add AbstractNetworkExploration to replace previous NetworkExploration
Fix TreeExploration and WebExploration
